requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
dataclasses-json==0.6.7
fastapi==0.115.8
filetype==1.2.0
frozenlist==1.5.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.68.0
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.19
langchain-community==0.3.18
langchain-core==0.3.37
langchain-google-genai==2.0.10
langchain-text-splitters==0.3.6
langgraph==0.2.74
langgraph-checkpoint==2.0.16
langgraph-sdk==0.1.53
langsmith==0.3.10
marshmallow==3.26.1
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
numpy==2.2.3
orjson==3.10.15
packaging==24.2
passlib==1.7.4
propcache==0.3.0
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic-settings==2.8.0
pydantic_core==2.27.2
pyparsing==3.2.1
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.45.3
tenacity==9.0.0
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
yarl==1.18.3
zstandard==0.23.0
